racehorse/requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==35.0.0
cssselect==1.1.0
fake-useragent==0.1.11
h11==0.12.0
html5lib==1.1
idna==3.3
importlib-metadata==4.10.0
lxml==4.7.1
outcome==1.1.0
parse==1.19.0
pdfminer.six==20211012
pycparser==2.21
pyee==8.2.2
pyOpenSSL==21.0.0
PyPDF2==1.26.0
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
requests==2.27.1
requests-html==0.10.0
selenium==4.1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.9
w3lib==1.22.0
webencodings==0.5.1
websockets==9.1
wsproto==1.0.0
zipp==3.7.0
